+++ egglog/src/ast/mod.rs
@@ -2,7 +2,6 @@ pub mod check_shadowing;
 pub mod desugar;
 mod expr;
 mod parse;
-pub mod proof_global_remover;
 pub mod remove_globals;
 
 use std::cmp::max;